requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.24
botocore==1.29.24
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.3
Pillow==9.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
